Move wake_signal under new util module
Introduce a top-level `util` module to hold generic helpers, and move
`wake_signal` into it. Updates the two callers in `memory::limiter`
and `memory::maintenance` to the new path.

Addresses PR feedback that `wake_signal` should sit in a dedicated
`util` module so other small generic types can join it later.
