+++ mountpoint-s3-fs/src/lib.rs
@@ -21,7 +21,7 @@ pub mod s3;
 mod superblock;
 mod sync;
 pub mod upload;
-mod wake_signal;
+mod util;
 
 pub use async_util::Runtime;
 pub use config::MountpointConfig;
