+++ mountpoint-s3-fs/src/util.rs
@@ -1,0 +1,3 @@
+//! Generic utility types reused across the crate.
+
+pub mod wake_signal;
